+++ test-parts/part-b-external_modules/modules/pkg-validator.js
@@ -1,19 +1,16 @@
 const chai = require("chai");
 const expect = require("chai").expect;
-const chaiThings = require('chai-things');
-const sinon = require('sinon');
 
 const commonPaths = require("../../../app/paths/files/app-paths");
 const commonFunctionsFile = require(commonPaths.testCommonFull);
 const ipStrings = require("../sub-modules/ip-strings");
-const validatorModule = require("validator");
+const validator = require("validator");
 
 function testValidatorDependency()
 {
 	describe("Validator", function()
 	{
-		verifyValidatorExists();
-		verifyIpFunction();
+		verifyFunctionsExist();
 		verifyIpAddressVerFour();
 		verifyIpAddressVerSix();
 		verifyIpAddressInvalid();
@@ -23,28 +20,22 @@ function testValidatorDependency()
 	
 }
 
-function verifyValidatorExists()
-{
-	describe("Library", function()
-	{
-		it("Exists", function()
-		{
-			commonFunctionsFile.testPresent(validatorModule);
-			expect(validatorModule).to.be.an("object");
-		});
-	});
-}
 
-function verifyIpFunction()
+
+function verifyFunctionsExist()
 {
-	var ipSpy = null;
-	
-	describe("IP Address (isIP)", function()
+	describe("Functions Exist", function()
 	{
-		it("Function Exists", function()
+		it("IP Address (isIP)", function()
 		{
-			commonFunctionsFile.testObjectPropertyDefinition(validatorModule, 'isIP');
-			commonFunctionsFile.testObjectPropertyContent(validatorModule, 'isIP', 'function');
+			commonFunctionsFile.testObjectPropertyDefinition(validator, 'isIP');
+			commonFunctionsFile.testObjectPropertyContent(validator, 'isIP', 'function');
+		});
+		
+		it("MAC Address (isMACAddress)", function()
+		{
+			commonFunctionsFile.testObjectPropertyDefinition(validator, 'isMACAddress');
+			commonFunctionsFile.testObjectPropertyContent(validator, 'isMACAddress', 'function');
 		});
 		
 	});
@@ -53,57 +44,30 @@ function verifyIpFunction()
 
 function verifyIpAddressVerFour()
 {
-	var ipSpy = null;
+	var validResult = false;
+	var emptyResult = false;
+	var invalidResult = false;
 	
 	describe("IP Address - Version 4", function()
 	{
-		it("Spy Established", function(done)
+		
+		it("Valid Format", function()
 		{
-			ipSpy = sinon.spy(validatorModule, 'isIP');
-			done();
+			validResult = validator.isIP(ipStrings.testString, 4);
+			expect(validResult).to.be.true;
 		});
 		
-		it("Call - Valid", function()
+		it("Empty", function()
 		{
-			validatorModule.isIP(ipStrings.testString, 4);
-			
-			expect(ipSpy.calledOnce).to.be.true;
-			expect(ipSpy.firstCall.args).to.deep.equal([ipStrings.testString, 4]);
-			expect(ipSpy.firstCall.exception).to.be.undefined;
-			
-			
-			commonFunctionsFile.testPresent(ipSpy.firstCall.returnValue);
-			expect(ipSpy.firstCall.returnValue).to.be.true;
-		});
-		
-		it("Call - Empty", function()
-		{
-			validatorModule.isIP("", 4);
-			
-			expect(ipSpy.calledTwice).to.be.true;
-			expect(ipSpy.secondCall.args).to.deep.equal(["", 4]);
-			expect(ipSpy.secondCall.exception).to.be.undefined;
-			
-			commonFunctionsFile.testPresent(ipSpy.secondCall.returnValue);
-			expect(ipSpy.secondCall.returnValue).to.be.false;
+			emptyResult = validator.isIP("", 4);
+			expect(emptyResult).to.be.false;
 		});
 		
 		
-		it("Call - Invalid Format", function()
+		it("Invalid Format", function()
 		{
-			validatorModule.isIP(ipStrings.invalidString, 4);
-			
-			expect(ipSpy.called).to.be.true;
-			expect(ipSpy.lastCall.args).to.deep.equal([ipStrings.invalidString, 4]);
-			expect(ipSpy.secondCall.exception).to.be.undefined;
-			
-			commonFunctionsFile.testPresent(ipSpy.lastCall.returnValue);
-			expect(ipSpy.lastCall.returnValue).to.be.false;
-		});
-		
-		it("Complete", function()
-		{
-			ipSpy.restore();
+			invalidResult = validator.isIP(ipStrings.invalidString, 4);
+			expect(invalidResult).to.be.false;
 		});
 		
 	});
@@ -113,59 +77,40 @@ function verifyIpAddressVerFour()
 
 function verifyIpAddressVerSix()
 {
-	var sixValid = "1234:5678:9ABC:DEF:0000:0000:0000:0001";
-	var sixLoop = "::1";
-	var sixInvalid = "1234:5678:9101112:13141516:ABCD:EFGH:IJKL:MNOP";
-	var ipSpy = null;
+	var validString = "1234:5678:9ABC:DEF:0000:0000:0000:0001";
+	var loopbackString = "::1";
+	var invalidString = "1234:5678:9101112:13141516:ABCD:EFGH:IJKL:MNOP";
+	
+	var validResult = false;
+	var loopbackResult = false;
+	var emptyResult = false;
+	var invalidResult = false;
 	
 	describe("IP Address - Version 6", function()
-	{
-		it("Spy Established", function(done)
+	{	
+		it("Valid Format", function()
 		{
-			ipSpy = sinon.spy(validatorModule, 'isIP');
-			done();
+			validResult = validator.isIP(validString, 6);
+			expect(validResult).to.be.true;
 		});
 		
-		it("Call - Valid", function()
+		it("Loopback", function()
 		{
-			validatorModule.isIP(sixValid, 6);
-			
-			expect(ipSpy.calledOnce).to.be.true;
-			expect(ipSpy.firstCall.args).to.deep.equal([sixValid, 6]);
-			expect(ipSpy.firstCall.exception).to.be.undefined;
-			
-			commonFunctionsFile.testPresent(ipSpy.firstCall.returnValue);
-			expect(ipSpy.firstCall.returnValue).to.be.true;
-			
+			loopbackResult = validator.isIP(loopbackString, 6);
+			expect(loopbackResult).to.be.true;
 		});
 		
-		it("Call - Loopback", function()
+		it("Empty", function()
 		{
-			validatorModule.isIP(sixLoop, 6);
-			
-			expect(ipSpy.calledTwice).to.be.true;
-			expect(ipSpy.secondCall.args).to.deep.equal([sixLoop, 6]);
-			expect(ipSpy.secondCall.exception).to.be.undefined;
-			
-			commonFunctionsFile.testPresent(ipSpy.secondCall.returnValue);
-			expect(ipSpy.secondCall.returnValue).to.be.true;
+			emptyResult = validator.isIP("", 6);
+			expect(emptyResult).to.be.false;
 		});
 		
-		it("Call - Invalid", function()
-		{
-			validatorModule.isIP(sixInvalid, 6);
-			
-			expect(ipSpy.called).to.be.true;
-			expect(ipSpy.lastCall.args).to.deep.equal([sixInvalid, 6]);
-			expect(ipSpy.lastCall.exception).to.be.undefined;
-			
-			commonFunctionsFile.testPresent(ipSpy.lastCall.returnValue);
-			expect(ipSpy.lastCall.returnValue).to.be.false;
-		});
 		
-		it("Complete", function()
+		it("Invalid Format", function()
 		{
-			ipSpy.restore();
+			invalidResult = validator.isIP(invalidString, 6);
+			expect(invalidResult).to.be.false;
 		});
 		
 	});
@@ -176,16 +121,16 @@ function verifyIpAddressInvalid()
 {
 	describe("IP Address - Invalid", function()
 	{
-		it("Call - Invalid Type", function()
+		it("Invalid Type", function()
 		{
-			var iRes = ipInvalidCall(-1);
-			commonFunctionsFile.testInvalidFunctionResult(iRes, "Expected a string but received a number");
+			var invalidTypeRes = ipInvalidCall(-1);
+			commonFunctionsFile.testInvalidFunctionResult(invalidTypeRes, "Expected a string but received a number");
 		});
 		
-		it("Call - Null", function()
+		it("Null", function()
 		{
-			var nRes = ipInvalidCall(null);
-			commonFunctionsFile.testInvalidFunctionResult(nRes, "Expected a string but received a null");
+			var nullRes = ipInvalidCall(null);
+			commonFunctionsFile.testInvalidFunctionResult(nullRes, "Expected a string but received a null");
 		});
 	});
 }
@@ -196,23 +141,24 @@ function verifyMacAddressFunction()
 	var macValid = "00-14-22-01-23-45";
 	var macInvalid = "Invalid MAC";
 	
-	describe("MAC Address (isMACAddress)", function()
+	describe("MAC Address", function()
 	{
-		it("Function Exists", function()
+		it("Valid Format", function()
 		{
-			commonFunctionsFile.testObjectPropertyDefinition(validatorModule, 'isMACAddress');
-			commonFunctionsFile.testObjectPropertyContent(validatorModule, 'isMACAddress', 'function');
-		});
-		
-		it("Call - Valid Format", function()
-		{
-			var validResult = validatorModule.isMACAddress(macValid);
+			var validResult = validator.isMACAddress(macValid);
 			expect(validResult).to.be.true;
 		});
 		
-		it("Call - Invalid Format", function()
+		it("Empty", function()
 		{
-			var invalidResult = validatorModule.isMACAddress(macInvalid);
+			var emptyResult = validator.isMACAddress("");
+			expect(emptyResult).to.be.false;
+		});
+		
+		
+		it("Invalid Format", function()
+		{
+			var invalidResult = validator.isMACAddress(macInvalid);
 			expect(invalidResult).to.be.false;
 		});
 		
@@ -223,13 +169,12 @@ function verifyMacAddressFunction()
 
 function ipInvalidCall(ipArg)
 {
-	var funcRes = null;
 	var validationComplete = false;
 	var validationMessage = "";
 	
 	try
 	{
-		funcRes = validatorModule.isIP(ipArg);
+		validator.isIP(ipArg);
 		validationComplete = true;
 	}
 	catch(e)